Reorganised and renamed files.
