+++ src/Components/My_NavBar.js
@@ -1,5 +1,5 @@
 import React, { Component } from 'react';
-import './App.css';
+import '../App.css';
 
 class My_NavBar extends Component{
     render(){
