+++ src/Screens/Home_Page.js
@@ -1,9 +1,8 @@
 import React, { Component } from 'react';
-import logo from './logo.svg';
-import './App.css';
-import My_Header from './My_Header';
-import My_Table from './My_Table';
-import My_NavBar from './My_NavBar';
+import '../App.css';
+import My_Header from '../Components/My_Header';
+import My_Table from '../Components/My_Table';
+import My_NavBar from '../Components/My_NavBar';
 
 class App extends Component {
   render() {
